requirements.txt
python-dotenv
pandas
openpyxl
pdfplumber
python-docx
